refactor: 환자의 안전구역 내 존재 확인 api 추가
## 작업 내용
- 안전구역 내 존재 확인 API 개발
기존: 소켓통신을 통해 안전구역 내 존재 확인
변경: 환자 번호, 위도, 경도를 통해 안전구역 내 존재 확인 API 개발
- 사용하지 않는 코드 정리
