+++ src/main/java/kr/co/onehunnit/onhunnit/domain/diary/Diary.java
@@ -3,6 +3,7 @@ package kr.co.onehunnit.onhunnit.domain.diary;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -35,7 +36,7 @@ public class Diary extends BaseTimeEntity {
 
 	private boolean shared;
 
-	@ManyToOne
+	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "patient_id")
 	private Patient patient;
