+++ src/main/java/kr/co/onehunnit/onhunnit/domain/patient_Caregiver/PatientCaregiver.java
@@ -2,12 +2,13 @@ package kr.co.onehunnit.onhunnit.domain.patient_Caregiver;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import kr.co.onehunnit.onhunnit.domain.account.Caregiver;
+import kr.co.onehunnit.onhunnit.domain.caregiver.Caregiver;
 import kr.co.onehunnit.onhunnit.domain.global.BaseTimeEntity;
 import kr.co.onehunnit.onhunnit.domain.patient.Patient;
 import lombok.AllArgsConstructor;
@@ -29,11 +30,11 @@ public class PatientCaregiver extends BaseTimeEntity {
 
 	private boolean is_accepted;
 
-	@ManyToOne
+	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "patient_id")
 	private Patient patient;
 
-	@ManyToOne
+	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "caregiver_id")
 	private Caregiver caregiver;
